SIGCONT not trappable in VMS.
p4raw-id: //depot/perl@10937

+++ ext/POSIX/sigaction.t
@@ -117,13 +117,17 @@ eval {
 };
 print $@ ? "ok 17\n" : "not ok 17\n";
 
-$newaction=POSIX::SigAction->new(sub { $ok10=1; });
-if (eval { SIGCONT; 1 }) {
-    sigaction(SIGCONT, POSIX::SigAction->new('DEFAULT'));
-    {
-	local($^W)=0;
-	kill 'CONT', $$;
+if ($^O e 'VMS') {
+    print "ok 18 # Skip: SIGCONT not trappable in $^O\n";
+} else {
+    $newaction=POSIX::SigAction->new(sub { $ok10=1; });
+    if (eval { SIGCONT; 1 }) {
+	sigaction(SIGCONT, POSIX::SigAction->new('DEFAULT'));
+	{
+	    local($^W)=0;
+	    kill 'CONT', $$;
+	}
     }
+    print $bad18 ? "not ok 18\n" : "ok 18\n";
 }
-print $bad18 ? "not ok 18\n" : "ok 18\n";
